chore:获取的size进行单位转换

+++ android/src/main/java/im/shimo/react/keyboard/KeyboardModule.java
@@ -52,7 +52,7 @@ public class KeyboardModule extends ReactContextBaseJavaModule {
 
     @ReactMethod
     public void getNavigationSize(Promise promise) {
-        int size = KeyboardViewManager.getNavigationSize();
+        float size = KeyboardViewManager.getNavigationSize();
         promise.resolve(size);
     }
 }
